refactor(chat): メッセージパーツの統合と不要ファイルの削除
