+++ apps/web/src/features/chat/components/chat.tsx
@@ -18,8 +18,7 @@ import { useChatAttachment } from '../hooks/use-chat-attachment'
 import { formatBytes } from '../utils/format-bytes'
 import { loadChatAttachmentFromFileSystem } from '../utils/load-chat-attachment-from-file-system'
 import { useChatSession } from './chat-session-provider'
-import { MessageBody, MessageHeader, Messages } from './message'
-import { MessageParts } from './parts/message-part'
+import { MessageBody, MessageHeader, MessageParts, Messages } from './message'
 import { NotImplementedPartContent } from './parts/not-implemented-part-content'
 import { TextContent } from './parts/text-content'
 import { ThinkingContent } from './parts/thinking-content'
